+++ coreblocks/cache/icache.py
@@ -332,4 +332,3 @@ class ICacheMemory(Elaboratable):
             ]
 
         return m
-
